split av/errors.rs & add some wrap **NO TEST**
