+++ src/core/errors.rs
@@ -1,5 +1,3 @@
-
-
 #[repr(C)]
 #[derive(Clone, Copy, PartialEq, Eq, Debug)]
 pub enum InitError {
